♻️ :: 단일 화면 위젯 파일을 해당 화면으로 통합
여러 화면에서 공유되지 않는 위젯 파일(outing/profile/member/report)을
사용하는 화면 파일 내부로 이동하여 불필요한 파일 분리 제거
