+++ lib/features/report/presentation/screens/admin_report_list_screen.dart
@@ -10,9 +10,10 @@ import 'package:goms/core/widgets/bottom_sheets/filter_button.dart';
 import 'package:goms/core/widgets/buttons/qr_button.dart';
 import 'package:goms/core/widgets/scaffolds/base_scaffold.dart';
 import 'package:goms/features/map/review/domain/enums/report_status.dart';
-import 'package:goms/features/report/presentation/models/report_summary_model.dart';
+import 'package:goms/features/report/domain/entities/report_summary_entity.dart';
 import 'package:goms/features/report/presentation/providers/admin_report_providers.dart';
-import 'package:goms/features/report/presentation/widgets/report_filter_bottom_sheet.dart';
+import 'package:goms/core/widgets/bottom_sheets/common_bottom_sheet.dart';
+import 'package:goms/core/widgets/chips/category_chip.dart';
 import 'package:intl/intl.dart';
 
 final _reportSearchQueryProvider = NotifierProvider.autoDispose
@@ -160,17 +161,17 @@ class _ReportListBody extends ConsumerWidget {
     required this.reportStatusFilter,
   });
 
-  final AsyncValue<List<ReportSummaryModel>> pendingReportsAsync;
-  final AsyncValue<List<ReportSummaryModel>> resolvedReportsAsync;
+  final AsyncValue<List<ReportSummaryEntity>> pendingReportsAsync;
+  final AsyncValue<List<ReportSummaryEntity>> resolvedReportsAsync;
   final String query;
   final ReportStatus? reportStatusFilter;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final pendingReports =
-        pendingReportsAsync.asData?.value ?? const <ReportSummaryModel>[];
+        pendingReportsAsync.asData?.value ?? const <ReportSummaryEntity>[];
     final resolvedReports =
-        resolvedReportsAsync.asData?.value ?? const <ReportSummaryModel>[];
+        resolvedReportsAsync.asData?.value ?? const <ReportSummaryEntity>[];
 
     final hasAnyData =
         pendingReportsAsync.hasValue || resolvedReportsAsync.hasValue;
@@ -203,7 +204,7 @@ class _ReportListBody extends ConsumerWidget {
       );
     }
 
-    final reports = <ReportSummaryModel>[...pendingReports, ...resolvedReports]
+    final reports = <ReportSummaryEntity>[...pendingReports, ...resolvedReports]
       ..sort((a, b) {
         final aTime =
             a.reportCreatedAt ?? DateTime.fromMillisecondsSinceEpoch(0);
@@ -213,7 +214,7 @@ class _ReportListBody extends ConsumerWidget {
       });
 
     final normalizedQuery = query.trim().toLowerCase();
-    final List<ReportSummaryModel> filteredReports = reports.where((report) {
+    final List<ReportSummaryEntity> filteredReports = reports.where((report) {
       final matchesStatus = matchesReportStatusFilter(
         filter: reportStatusFilter,
         status: report.reportStatus,
@@ -314,7 +315,7 @@ class _ReportListTile extends StatelessWidget {
     required this.onTap,
   });
 
-  final ReportSummaryModel report;
+  final ReportSummaryEntity report;
   final VoidCallback onTap;
 
   @override
@@ -373,7 +374,7 @@ class _ReportListTile extends StatelessWidget {
   }
 }
 
-String _reportHeadline(ReportSummaryModel report) {
+String _reportHeadline(ReportSummaryEntity report) {
   final reportContent = report.reportContent?.trim();
   if (reportContent != null && reportContent.isNotEmpty) {
     return reportContent;
@@ -381,7 +382,7 @@ String _reportHeadline(ReportSummaryModel report) {
   return '신고 내용 없음';
 }
 
-String _reportPlaceName(ReportSummaryModel report) {
+String _reportPlaceName(ReportSummaryEntity report) {
   final placeName = report.placeName?.trim();
   final placeAddress = report.placeAddress?.trim();
 
@@ -458,5 +459,166 @@ class _ReportErrorView extends StatelessWidget {
         ),
       ],
     );
+  }
+}
+
+// ---------------------------------------------------------------------------
+// ReportFilterBottomSheet (moved from widgets/report_filter_bottom_sheet.dart)
+// ---------------------------------------------------------------------------
+
+class ReportFilterSelection {
+  const ReportFilterSelection({
+    this.reportStatus,
+  });
+
+  final ReportStatus? reportStatus;
+
+  ReportFilterSelection copyWith({
+    ReportStatus? reportStatus,
+    bool clearReportStatus = false,
+  }) {
+    return ReportFilterSelection(
+      reportStatus:
+          clearReportStatus ? null : (reportStatus ?? this.reportStatus),
+    );
+  }
+}
+
+final _reportFilterSelectionProvider = NotifierProvider.autoDispose.family<
+    _ReportFilterSelectionNotifier,
+    ReportFilterSelection,
+    (Object, ReportFilterSelection)>(_ReportFilterSelectionNotifier.new);
+
+class _ReportFilterSelectionNotifier extends Notifier<ReportFilterSelection> {
+  _ReportFilterSelectionNotifier(this.args);
+
+  final (Object, ReportFilterSelection) args;
+
+  @override
+  ReportFilterSelection build() => args.$2;
+
+  void setSelection(ReportFilterSelection selection) => state = selection;
+}
+
+class ReportFilterBottomSheet extends ConsumerStatefulWidget {
+  const ReportFilterBottomSheet({
+    super.key,
+    this.initialSelection = const ReportFilterSelection(),
+    this.onApply,
+    this.onReset,
+  });
+
+  final ReportFilterSelection initialSelection;
+  final ValueChanged<ReportFilterSelection>? onApply;
+  final VoidCallback? onReset;
+
+  @override
+  ConsumerState<ReportFilterBottomSheet> createState() =>
+      _ReportFilterBottomSheetState();
+}
+
+class _ReportFilterBottomSheetState
+    extends ConsumerState<ReportFilterBottomSheet> {
+  late final Object _providerIdentity;
+
+  (Object, ReportFilterSelection) get _providerKey =>
+      (_providerIdentity, widget.initialSelection);
+
+  @override
+  void initState() {
+    super.initState();
+    _providerIdentity = Object();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final selection = ref.watch(_reportFilterSelectionProvider(_providerKey));
+
+    return CommonBottomSheet(
+      title: '필터',
+      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 24),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            '상태',
+            style: AppTextStyles.title3.copyWith(
+              color: context.isLightMode ? Colors.black : Colors.white,
+            ),
+          ),
+          AppGap.v12,
+          Row(
+            children: [
+              Expanded(
+                child: CategoryChip(
+                  category: '처리전',
+                  selected: selection.reportStatus == ReportStatus.pending,
+                  textColor: context.mainTextColor,
+                  onChanged: (selected) => _updateSelection(
+                    selection.copyWith(
+                      reportStatus: selected ? ReportStatus.pending : null,
+                      clearReportStatus: !selected,
+                    ),
+                  ),
+                ),
+              ),
+              AppGap.h8,
+              Expanded(
+                child: CategoryChip(
+                  category: '처리완료',
+                  selected: selection.reportStatus == ReportStatus.approved,
+                  textColor: context.mainTextColor,
+                  onChanged: (selected) => _updateSelection(
+                    selection.copyWith(
+                      reportStatus: selected ? ReportStatus.approved : null,
+                      clearReportStatus: !selected,
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+          AppGap.v24,
+          GestureDetector(
+            onTap: _handleReset,
+            child: Container(
+              height: 44,
+              width: double.infinity,
+              decoration: BoxDecoration(
+                color: AppColors.negative.withValues(alpha: 0.25),
+                borderRadius: BorderRadius.circular(8),
+              ),
+              alignment: Alignment.center,
+              child: Text(
+                '필터 초기화',
+                style: AppTextStyles.text2.copyWith(
+                  color: AppColors.negative,
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _handleReset() {
+    ref
+        .read(_reportFilterSelectionProvider(_providerKey).notifier)
+        .setSelection(const ReportFilterSelection());
+    widget.onReset?.call();
+    Navigator.pop(context);
+  }
+
+  void _updateSelection(ReportFilterSelection selection) {
+    ref
+        .read(_reportFilterSelectionProvider(_providerKey).notifier)
+        .setSelection(selection);
+    _notifySelectionChanged(selection);
+  }
+
+  void _notifySelectionChanged(ReportFilterSelection selection) {
+    widget.onApply?.call(selection);
   }
 }
